fix: BookStatus 를 Book.java 밖으로 빼고 관련된 파일 처리
