+++ src/main/java/com/libraryManagement/domain/Book.java
@@ -1,6 +1,6 @@
 package com.libraryManagement.domain;
 
-import static com.libraryManagement.domain.Book.BookStatus.*;
+import static com.libraryManagement.domain.BookStatus.*;
 
 public class Book {
     private final long id;
@@ -111,22 +111,5 @@ public class Book {
                 "\n상태 : " + status + "\n";
     }
 
-    public enum BookStatus {
-
-        AVAILABLE("대여가능"),
-        RENT("대여중"),
-        ORGANIZING("정리중"),
-        LOST("분실됨"),
-        DELETE("삭제됨");
-
-        final private String name;
-
-        public String getName() {
-            return name;
-        }
-
-        BookStatus(String name) {
-            this.name = name;
-        }
-    }
 }
+
