+++ src/main/java/com/libraryManagement/domain/BookStatus.java
@@ -1,0 +1,20 @@
+package com.libraryManagement.domain;
+
+public enum BookStatus {
+
+    AVAILABLE("대여가능"),
+    RENT("대여중"),
+    ORGANIZING("정리중"),
+    LOST("분실됨"),
+    DELETE("삭제됨");
+
+    final private String name;
+
+    public String getName() {
+        return name;
+    }
+
+    BookStatus(String name) {
+        this.name = name;
+    }
+}
